feat: Add custom picker and change styling

+++ PennyWise/Models/Category.swift
@@ -11,6 +11,6 @@ struct Category: Identifiable {
     let id = UUID()
     let name: String
     let iconName: String
-    let budget: Int
-    let remaining: Int
+    let budget: Double
+    let remaining: Double
 }
